dialogue batching refactor (cond)
